chore: moved fn into packet.go and added no-op telemetry v2 functions
